totalSentiment() method
This method loops through a review, and adds up the sentiment value of all of the words within the review, based on their sentiment value in the spreadsheet document. I used two variables, an empty String to collect each word as the program loops through the review, and a double value to hold the total sentiment value; each word's value is added to this variable after the completion of the loop through said word.