fakeReview() method
This method replaces all the adjectives in "Simple review" that have a "*" in front of it with a random adjective. It loops through the review, and when it reaches wither a space or the last word in the review, then it adds the last character to the placeHolder variable. If the word starts with a "*," then it generates a new adjective that will take the spot in the review of the old adjective with a "*" in front of it. Finally, it sets the placeHolder variable equal to an empty string, so that this variable can be used again and again, as intended, in this loop. Once the loop finishes iterating, then it will return the new fake review with fake adjectives.