fakeReviewStronger()
This method is intended to loop through the review, and replace any adjectives with a "*" at the front of the word with an adjective that has a stronger (more positive or more negative depending on the word) sentiment value. For example, it may change the word "like" to the word "love," because it has a stronger positive sentiment value. Since the first if statement in the method checks for the "*," then I don't have to worry about it anymore throughout the program.