update fakeReview() method
This method simply replaces each adjective that has a "*" before it with a different random adjective. This method loops through the review until the current character is a "*." When the current character is NOT a "*," then it will add the current character to the empty string variable that is used to hold the current adjective.  If the current character is a "*" then it will loop through the word that follows the "*." Finally, it will add a random adjective (generated by the randomAdjective() method) to the string that holds the new, fake review.